scripts/requirements.txt
# Microsoft Graph API Scripts Dependencies

# Microsoft Authentication Library
msal>=1.24.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For interactive notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
